Upload SDE_Appointment
The point of this assignment was for me to get familiar with the concept of apportionment, which is a democratic process used to divide indivisible resources — like seats in the U.S. House of Representatives — based on state populations. By implementing Alexander Hamilton's apportionment algorithm, I got to practice my Java programming skills, collaborate with others using Git and GitHub, and learn how to build software iteratively while responding to changing requirements. This assignment also helped me get comfortable working in a team and using tools like Gradle.